requirements/pip.txt
asgiref==3.2.5
astroid==2.3.3
autopep8==1.5
backcall==0.1.0
certifi==2020.4.5.1
chardet==3.0.4
decorator==4.4.2
Django==3.0.4
djangorestframework==3.11.0
httpie==2.0.0
idna==2.9
ipdb==0.13.2
ipython==7.13.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.16.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
ptyprocess==0.6.0
pycodestyle==2.5.0
Pygments==2.6.1
pylint==2.4.4
pylint-django==2.0.14
pylint-plugin-utils==0.6
PyMySQL==0.9.3
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
traitlets==4.3.3
typed-ast==1.4.1
urllib3==1.25.8
wcwidth==0.1.9
wrapt==1.11.2
